requirements.txt
flask
flask-cors
feedparser
chromadb
ollama
openai
python-dotenv
